cv_business_slides/requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
matplotlib>=3.4.0
Pillow>=8.0.0
torch>=1.9.0
torchvision>=0.10.0
jupyter>=1.0.0 